+++ server/api.js
@@ -28,16 +28,24 @@ router.use(middleware.authSanitizer);
 
 
 /* Auth */
-router.route('/v1/auth/login').post(authAPI.validate('login'), middleware.checkValidationErrors, authAPI.login);
+router.route('/v1/auth/login').post(authAPI.validate('login'),
+    middleware.checkValidationErrors, authAPI.login);
 
 
 /* Organizations */
-router.route('/v1/org/info').get(orgAPI.validate('getinfo'), middleware.checkValidationErrors, orgAPI.getOrganizationInfo);
+router.route('/v1/org/info').get(orgAPI.validate('getinfo'),
+    middleware.checkValidationErrors, orgAPI.getOrganizationInfo);
 
 
-/* Adoption Reports (can be anonymous) */
-router.route('/v1/adoptionreport').post(middleware.checkLibreCommons, adoptionReportAPI.validate('submitReport'),
+/* Adoption Reports */
+// (submission route can be anonymous)
+router.route('/v1/adoptionreport').post(middleware.checkLibreCommons,
+    adoptionReportAPI.validate('submitReport'),
     middleware.checkValidationErrors, adoptionReportAPI.submitReport);
+router.route('/v1/adoptionreports').get(middleware.checkLibreCommons,
+    authAPI.verifyRequest,
+    adoptionReportAPI.validate('getReports'), middleware.checkValidationErrors,
+    adoptionReportAPI.getReports); // TODO: Restrict route access
 
 /* Search */
 //router.route('/v1/search').get(authAPI.verifyRequest, searchAPI.performSearch);
@@ -45,55 +53,88 @@ router.route('/v1/adoptionreport').post(middleware.checkLibreCommons, adoptionRe
 
 /* Users */
 //router.route('/v1/user/createuserbasic').post(userAPI.createUserBasic);
-router.route('/v1/user/basicinfo').get(authAPI.verifyRequest, userAPI.basicUserInfo);
-router.route('/v1/user/accountinfo').get(authAPI.verifyRequest, userAPI.basicAccountInfo);
-router.route('/v1/user/getadmins').get(authAPI.verifyRequest, userAPI.getAdmins);
-router.route('/v1/user/getdevelopers').get(authAPI.verifyRequest, userAPI.getDevelopers);
-router.route('/v1/user/getharvesters').get(authAPI.verifyRequest, userAPI.getHarvesters);
+router.route('/v1/user/basicinfo').get(authAPI.verifyRequest,
+    userAPI.basicUserInfo);
+router.route('/v1/user/accountinfo').get(authAPI.verifyRequest,
+    userAPI.basicAccountInfo);
+router.route('/v1/user/getadmins').get(authAPI.verifyRequest,
+    userAPI.getAdmins);
+router.route('/v1/user/getdevelopers').get(authAPI.verifyRequest,
+    userAPI.getDevelopers);
+router.route('/v1/user/getharvesters').get(authAPI.verifyRequest,
+    userAPI.getHarvesters);
 
 
 
 /* Announcements */
-router.route('/v1/announcements/create').post(authAPI.verifyRequest, announcementAPI.postAnnouncement);
-router.route('/v1/announcements/all').get(authAPI.verifyRequest, announcementAPI.getAllAnnouncements);
-router.route('/v1/announcements/recent').get(authAPI.verifyRequest, announcementAPI.getRecentAnnouncement);
+router.route('/v1/announcements/create').post(authAPI.verifyRequest,
+    announcementAPI.postAnnouncement);
+router.route('/v1/announcements/all').get(authAPI.verifyRequest,
+    announcementAPI.getAllAnnouncements);
+router.route('/v1/announcements/recent').get(authAPI.verifyRequest,
+    announcementAPI.getRecentAnnouncement);
 
 
 /* Projects (General) */
-router.route('/v1/projects/all').get(authAPI.verifyRequest, sharedProjectsAPI.getAllUserProjects);
-router.route('/v1/projects/recent').get(authAPI.verifyRequest, sharedProjectsAPI.getRecentUserProjects);
+router.route('/v1/projects/all').get(authAPI.verifyRequest,
+    sharedProjectsAPI.getAllUserProjects);
+router.route('/v1/projects/recent').get(authAPI.verifyRequest,
+    sharedProjectsAPI.getRecentUserProjects);
 
 
 /* Harvesting */
 
 // Requests (LibreCommons) (can be anonymous)
-router.route('/v1/harvesting/request/new').post(middleware.checkLibreCommons, harvestingRequestsAPI.validate('addRequest'),
+router.route('/v1/harvesting/request/new').post(middleware.checkLibreCommons,
+    harvestingRequestsAPI.validate('addRequest'),
     middleware.checkValidationErrors, harvestingRequestsAPI.addRequest);
 
 // Targetlist
-router.route('/v1/harvesting/targetlist/all').get(authAPI.verifyRequest, harvestingTargetsAPI.getAllTargets);
-router.route('/v1/harvesting/targetlist/add').post(authAPI.verifyRequest, harvestingTargetsAPI.addTarget);
-router.route('/v1/harvesting/targetlist/targets/detail').get(authAPI.verifyRequest, harvestingTargetsAPI.getTargetDetail);
-router.route('/v1/harvesting/targetlist/targets/update').post(authAPI.verifyRequest, harvestingTargetsAPI.updateTarget);
-router.route('/v1/harvesting/targetlist/targets/delete').post(authAPI.verifyRequest, harvestingTargetsAPI.deleteTarget);
+router.route('/v1/harvesting/targetlist/all').get(authAPI.verifyRequest,
+    harvestingTargetsAPI.getAllTargets);
+router.route('/v1/harvesting/targetlist/add').post(authAPI.verifyRequest,
+    harvestingTargetsAPI.addTarget);
+router.route('/v1/harvesting/targetlist/targets/detail').get(
+    authAPI.verifyRequest, harvestingTargetsAPI.getTargetDetail);
+router.route('/v1/harvesting/targetlist/targets/update').post(
+    authAPI.verifyRequest, harvestingTargetsAPI.updateTarget);
+router.route('/v1/harvesting/targetlist/targets/delete').post(
+    authAPI.verifyRequest, harvestingTargetsAPI.deleteTarget);
 
 // Projects
-router.route('/v1/harvesting/projects/addexisting').post(authAPI.verifyRequest, harvestingProjectsAPI.addExistingProject);
-router.route('/v1/harvesting/projects/newfromtarget').post(authAPI.verifyRequest, harvestingProjectsAPI.newProjectFromTarget);
-router.route('/v1/harvesting/projects/newforassignee').post(authAPI.verifyRequest, harvestingProjectsAPI.newProjectForAssignee);
-router.route('/v1/harvesting/projects/detail').get(authAPI.verifyRequest, harvestingProjectsAPI.getProjectDetail);
-router.route('/v1/harvesting/projects/update').post(authAPI.verifyRequest, harvestingProjectsAPI.updateProject);
-router.route('/v1/harvesting/projects/flag').post(authAPI.verifyRequest, harvestingProjectsAPI.flagProject);
-router.route('/v1/harvesting/projects/unflag').post(authAPI.verifyRequest, harvestingProjectsAPI.unflagProject);
-router.route('/v1/harvesting/projects/markcompleted').post(authAPI.verifyRequest, harvestingProjectsAPI.markProjectCompleted);
-router.route('/v1/harvesting/projects/delete').post(authAPI.verifyRequest, harvestingProjectsAPI.deleteProject);
-router.route('/v1/harvesting/projects/current').get(authAPI.verifyRequest, harvestingProjectsAPI.getCurrentProjects);
-router.route('/v1/harvesting/projects/flagged').get(authAPI.verifyRequest, harvestingProjectsAPI.getFlaggedProjects);
-router.route('/v1/harvesting/projects/recentlycompleted').get(authAPI.verifyRequest, harvestingProjectsAPI.getRecentlyCompletedProjects);
-router.route('/v1/harvesting/projects/completed').get(authAPI.verifyRequest, harvestingProjectsAPI.getAllCompletedProjects);
-router.route('/v1/harvesting/projects/updates/all').get(authAPI.verifyRequest, harvestingProjectsAPI.getAllProgressUpdates);
-router.route('/v1/harvesting/projects/updates/new').post(authAPI.verifyRequest, harvestingProjectsAPI.addProgressUpdate);
-router.route('/v1/harvesting/projects/updates/delete').post(authAPI.verifyRequest, harvestingProjectsAPI.deleteProgressUpdate);
-router.route('/v1/harvesting/projects/updates/feed').get(authAPI.verifyRequest, harvestingProjectsAPI.getUpdatesFeed);
+router.route('/v1/harvesting/projects/addexisting').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.addExistingProject);
+router.route('/v1/harvesting/projects/newfromtarget').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.newProjectFromTarget);
+router.route('/v1/harvesting/projects/newforassignee').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.newProjectForAssignee);
+router.route('/v1/harvesting/projects/detail').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getProjectDetail);
+router.route('/v1/harvesting/projects/update').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.updateProject);
+router.route('/v1/harvesting/projects/flag').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.flagProject);
+router.route('/v1/harvesting/projects/unflag').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.unflagProject);
+router.route('/v1/harvesting/projects/markcompleted').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.markProjectCompleted);
+router.route('/v1/harvesting/projects/delete').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.deleteProject);
+router.route('/v1/harvesting/projects/current').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getCurrentProjects);
+router.route('/v1/harvesting/projects/flagged').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getFlaggedProjects);
+router.route('/v1/harvesting/projects/recentlycompleted').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getRecentlyCompletedProjects);
+router.route('/v1/harvesting/projects/completed').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getAllCompletedProjects);
+router.route('/v1/harvesting/projects/updates/all').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getAllProgressUpdates);
+router.route('/v1/harvesting/projects/updates/new').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.addProgressUpdate);
+router.route('/v1/harvesting/projects/updates/delete').post(
+    authAPI.verifyRequest, harvestingProjectsAPI.deleteProgressUpdate);
+router.route('/v1/harvesting/projects/updates/feed').get(
+    authAPI.verifyRequest, harvestingProjectsAPI.getUpdatesFeed);
 
 module.exports = router;